+++ src/MultiMapa.tpp
@@ -2,10 +2,10 @@
 
 // Busca binária genérica para inserção ordenada (para valores que têm operator< definido)
 template <typename Valor>
-static size_t lowerBoundValor(const Vector<Valor>& vec, const Valor& v) {
-    size_t l = 0, r = vec.get_size();
+static unsigned int lowerBoundValor(const Vector<Valor>& vec, const Valor& v) {
+    unsigned int l = 0, r = vec.get_size();
     while (l < r) {
-        size_t m = (l + r) / 2;
+        unsigned int m = (l + r) / 2;
         if (vec[m] < v) l = m + 1;
         else            r = m;
     }
@@ -14,10 +14,10 @@ static size_t lowerBoundValor(const Vector<Valor>& vec, const Valor& v) {
 
 // Especialização para ponteiros: compara conteúdo (*ptr)
 template <typename T>
-static size_t lowerBoundValor(const Vector<T*>& vec, T* const& v) {
-    size_t l = 0, r = vec.get_size();
+static unsigned int lowerBoundValor(const Vector<T*>& vec, T* const& v) {
+    unsigned int l = 0, r = vec.get_size();
     while (l < r) {
-        size_t m = (l + r) / 2;
+        unsigned int m = (l + r) / 2;
         if (*vec[m] < *v) l = m + 1;
         else              r = m;
     }
@@ -26,14 +26,14 @@ static size_t lowerBoundValor(const Vector<T*>& vec, T* const& v) {
 
 // Calcula índice de balde
 template <typename Chave, typename Valor>
-size_t MultiMapa<Chave,Valor>::indicePara(const Chave& chave) const {
+unsigned int MultiMapa<Chave,Valor>::indicePara(const Chave& chave) const {
     return hashFn(chave) % numBaldes;
 }
 
 // Construtor: inicializa 'numBaldes' baldes vazios
 template <typename Chave, typename Valor>
-MultiMapa<Chave,Valor>::MultiMapa(size_t numBaldes,
-                                   size_t (*hashFn)(const Chave&))
+MultiMapa<Chave,Valor>::MultiMapa(unsigned int numBaldes,
+                                   unsigned int (*hashFn)(const Chave&))
     : tabela(numBaldes), numBaldes(numBaldes), numPares(0), limiteCarga(0.75f), hashFn(hashFn) {
     // tabela já vem com "numBaldes" vetores vazios graças a Vector(size_t)
 }
@@ -44,13 +44,13 @@ void MultiMapa<Chave,Valor>::inserir(const Chave& chave, const Valor& valor) {
     if ((numPares + 1.0f) / numBaldes > limiteCarga)
         rehash();
 
-    size_t idx = indicePara(chave);
+    unsigned int idx = indicePara(chave);
     auto& balde = tabela[idx];
 
     // procura grupo existente
-    for (size_t g = 0; g < balde.get_size(); ++g) {
+    for (unsigned int g = 0; g < balde.get_size(); ++g) {
         if (balde[g].chave == chave) {
-            size_t pos = lowerBoundValor(balde[g].valores, valor);
+            unsigned int pos = lowerBoundValor(balde[g].valores, valor);
             balde[g].valores.insert(pos, valor);
             ++numPares;
             return;
@@ -78,12 +78,12 @@ Vector<Valor> MultiMapa<Chave,Valor>::obter(const Chave& chave) const {
 
 // Remove todo o grupo associado à chave e retorna quantos valores foram removidos
 template <typename Chave, typename Valor>
-size_t MultiMapa<Chave,Valor>::remover(const Chave& chave) {
-    size_t idx = indicePara(chave);
+unsigned int MultiMapa<Chave,Valor>::remover(const Chave& chave) {
+    unsigned int idx = indicePara(chave);
     auto& balde = tabela[idx];
-    for (size_t g = 0; g < balde.get_size(); ++g) {
+    for (unsigned int g = 0; g < balde.get_size(); ++g) {
         if (balde[g].chave == chave) {
-            size_t cnt = balde[g].valores.get_size();
+            unsigned int cnt = balde[g].valores.get_size();
             numPares -= cnt;
             balde.erase(g);
             return cnt;
@@ -95,18 +95,18 @@ size_t MultiMapa<Chave,Valor>::remover(const Chave& chave) {
 // Rehash: dobra número de baldes e reinsere grupos/valores
 template <typename Chave, typename Valor>
 void MultiMapa<Chave,Valor>::rehash() {
-    size_t novaCap = numBaldes * 2;
+    unsigned int novaCap = numBaldes * 2;
     Vector< Vector< GrupoChave<Chave,Valor> > > novaTabela(novaCap);
 
-    for (size_t b = 0; b < numBaldes; ++b) {
+    for (unsigned int b = 0; b < numBaldes; ++b) {
         auto& balde = tabela[b];
-        for (size_t g = 0; g < balde.get_size(); ++g) {
+        for (unsigned int g = 0; g < balde.get_size(); ++g) {
             const auto& grupo = balde[g];
-            size_t novoIdx = hashFn(grupo.chave) % novaCap;
+            unsigned int novoIdx = hashFn(grupo.chave) % novaCap;
             auto& destBalde = novaTabela[novoIdx];
             destBalde.push_back(GrupoChave<Chave,Valor>(grupo.chave));
             auto& destino = destBalde[destBalde.get_size() - 1];
-            for (size_t i = 0; i < grupo.valores.get_size(); ++i)
+            for (unsigned int i = 0; i < grupo.valores.get_size(); ++i)
                 destino.valores.push_back(grupo.valores[i]);
         }
     }
